requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
pytest-django
numpy
pandas
scikit-learn
joblib
nltk
django-cors-headers
django-filter
gunicorn
mongoengine
pymongo